Move method from trait to main class
